Remove MainTabView, replaced it with MyHabitsView as the main screen. Redesigned add habit button to match Reminders native app. Renamed StatisticsView to SettingsView. Updated MyHabitsView, NewHabitView, DatePickerView, and RoutineXApp. Refactored HabitViewModel and Habit model. Staying on the path to creating a truly minimalist and native iOS experience.

+++ RoutineX/Models/Habit.swift
@@ -13,7 +13,7 @@ struct Habit: Identifiable, Codable {
     var title: String
     var icon: String
     var color: String
-    var goal: Int
-    var progress: Int
+    var goal: Double
+    var progress: Double
     var createdAt: Date
 }
